please lint.
git-svn-id: file:///svn/unbound/trunk@3367 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ util/net_help.c
@@ -790,7 +790,7 @@ int ub_openssl_lock_init(void)
 #if defined(HAVE_SSL) && defined(OPENSSL_THREADS) && !defined(THREADS_DISABLED)
 	int i;
 	ub_openssl_locks = (lock_basic_t*)reallocarray(
-		NULL, CRYPTO_num_locks(), sizeof(lock_basic_t));
+		NULL, (size_t)CRYPTO_num_locks(), sizeof(lock_basic_t));
 	if(!ub_openssl_locks)
 		return 0;
 	for(i=0; i<CRYPTO_num_locks(); i++) {
